history update

+++ Models/Shikkhanobish/Subject.cs
@@ -12,7 +12,7 @@ namespace SHikkhanobishAPI.Models.Shikkhanobish
         public string title { get; set; }
         public string name { get; set; }
         public int tuitionRequest { get; set; }
-        public float avgRatting { get; set; }
+        public double avgRatting { get; set; }
         public int indexNo { get; set; }
         public string Response { get; set; }
     }
